+++ src/main/java/de/unisb/cs/st/evosuite/ga/stoppingconditions/MaxGenerationStoppingCondition.java
@@ -32,7 +32,7 @@ public class MaxGenerationStoppingCondition extends StoppingCondition {
 	private static final long serialVersionUID = 251196904115160351L;
 
 	/** Maximum number of iterations */
-	protected int max_iterations = Properties.GENERATIONS;
+	protected long max_iterations = Properties.GENERATIONS;
 
 	/** Maximum number of iterations */
 	protected int current_iteration = 0;
@@ -76,12 +76,12 @@ public class MaxGenerationStoppingCondition extends StoppingCondition {
 	 * @see de.unisb.cs.st.evosuite.ga.StoppingCondition#setLimit(int)
 	 */
 	@Override
-	public void setLimit(int limit) {
+	public void setLimit(long limit) {
 		max_iterations = limit;
 	}
 
 	@Override
-	public int getLimit() {
+	public long getLimit() {
 		return max_iterations;
 	}
 
@@ -89,7 +89,7 @@ public class MaxGenerationStoppingCondition extends StoppingCondition {
 	 * @see de.unisb.cs.st.evosuite.ga.StoppingCondition#getCurrentValue()
 	 */
 	@Override
-	public int getCurrentValue() {
+	public long getCurrentValue() {
 		return current_iteration;
 	}
